refactor(cli): extract session selection helpers
