+++ packages/agent-cli/agent-cli.cabal
@@ -53,7 +53,9 @@ library
                     , Agent.CLI.Provider.Switch
                     , Agent.CLI.Runtime
                     , Agent.CLI.Runtime.Types
+                    , Agent.CLI.Session.Choices
                     , Agent.CLI.Session.History
+                    , Agent.CLI.Session.Selection
                     , Agent.CLI.Session.Runtime.Types
                     , Agent.CLI.SessionAdmin
                     , Agent.CLI.StartupContext
